V4 Phase 3.5 tasks #32 + #6: BROADCAST_AUTH + app opcodes default-encrypted
Closes the last two pending Phase 3.5 work items, plus a pre-existing
flag-truncation bug and a dead-code regression introduced by task #47.

Task #32 — BROADCAST_AUTH for mesh broadcasts (auth-only HMAC):
- New constants in Wire.h: ESPNOW_V4_BROADCAST_AUTH_TAG_LEN (32) and
  ESPNOW_V4_MAX_BROADCAST_AUTHED_PLAINTEXT (186).
- v4_send_frame: if ESPNOW_V4_FLAG_BROADCAST_AUTH is set and the frame
  has a non-zero meshFingerprint with a derived group key, append an
  HMAC-SHA256 tag computed over (header[0..30] || payload) keyed by
  the mesh group key (esp-grup subkey from Phase 3.1).
- v4_broadcast: ORs in BROADCAST_AUTH on every broadcast send.
- v4_try_handle_incoming RX path: if flag set, look up mesh group key,
  verify HMAC over the same AAD, sodium_memcmp the tag, strip on
  success, silently drop on mismatch.
- Forged broadcasts from outside the mesh now drop at the auth layer.
- Payload stays plaintext per locked design decision (heartbeats,
  sensor data, device names already observable; auth is the win).

Task #6 — flip single-frame app opcodes to default-encrypted:
- Helper functions (v4_send_text, v4_send_user_sync,
  v4_send_command_response) now route through v4_send_encrypted_or_queue
  for single-frame sends. CMD_RESP > MAX_PLAINTEXT (202 B) still falls
  back to plaintext v4_send_chunked until task #51 (encrypted
  fragmentation) lands.
- Direct CMD-with-credentials TX sites flipped: roomcmd, tagcmd, all
  three espnow_remote variants (browse/fetch/generic). Payload format
  "user:pass:command" no longer goes plaintext on the air.
- METADATA_REQ and METADATA_PUSH (both forward + own) flipped.
- Sites that fail (no peer identity) return false; caller treats as
  "couldn't reach that peer." User runs espnowkeyex to establish
  identity, then retries.

Pre-existing bug fix: v4_send_frame's flags parameter was uint8_t but
EspNowV4Flags is a 16-bit enum. High-byte bits (BROADCAST_AUTH=0x0100,
SESSION_FRAME=0x0200, HANDSHAKE=0x0400) were silently truncated at
every callsite. Confirmed in earlier Phase 3.4 logs:
   "Type=13 MsgId=8 PayloadLen=126 Flags=0x01"
SESSION_OPEN was supposed to carry HANDSHAKE | ACK_REQ = 0x0401 but
emitted 0x01. The handshake worked because dispatch is by type, not by
flag — but the wire bit was lost. Widened uint8_t -> uint16_t in three
declarations (header + .cpp def + two extern decls).

Regression fix: METADATA_PUSH loop was gated on `peer.encrypt` (the
LMK flag). Task #47 ripped LMK and now always sets peer.encrypt=false,
so the metadata push loop never ran — silent dead code. New gate is
v4_send_encrypted_or_queue's internal "has peer identity?" check,
which is the right semantic anyway: only push metadata to peers we
have a long-term identity for.

+++ components/hardwareone/System_ESPNow.h
@@ -1113,8 +1113,11 @@ inline bool isBondModeOnline() { return false; }
 inline bool isBondSynced() { return false; }
 #endif // ENABLE_BONDED_MODE
 
-// V3 frame sending (for remote command execution from System_Utils)
-bool v4_send_frame(const uint8_t* dst, uint8_t type, uint8_t flags, uint32_t msgId,
+// V3 frame sending (for remote command execution from System_Utils).
+// Phase 3.5 task #32: flags widened uint8_t -> uint16_t so high-byte flag
+// bits (BROADCAST_AUTH/SESSION_FRAME/HANDSHAKE) survive instead of being
+// silently truncated.
+bool v4_send_frame(const uint8_t* dst, uint8_t type, uint16_t flags, uint32_t msgId,
                    const uint8_t* payload, uint16_t payloadLen, uint8_t ttl);
 
 // =============================================================================
